src/moonshot_utils/views.py
from rest_framework import status
from rest_framework.permissions import IsAuthenticatedOrReadOnly
from rest_framework.response import Response
from rest_framework.views import APIView


class MoonshotUtilsVersionView(APIView):
    permission_classes = [IsAuthenticatedOrReadOnly]

    def get(self, request):
        return Response(dict(version="0.2.0"), status=status.HTTP_200_OK)
